Add callback on event and authorization logic
This sets up a baseline to be able to have one door, opened by one
rolling code opener, that must present a code in the right range.  I
have only tested it with a Security 1.0 opener so far, and it does not
yet have code for actually opening a door via any mechanism.